requirements.txt
MetaTrader5
pandas
numpy
scikit-learn
onnx
onnxruntime
ta
flask
flask-socketio
websocket-client
python-socketio
eventlet
matplotlib
seaborn
plotly
requests
schedule
pytz
skl2onnx
google-cloud-run
google-cloud-storage